+++ components/02-elements/pager/pager.config.js
@@ -1,9 +1,10 @@
 module.exports = {
-  status: 'wip',
   collated: true,
   context: {
-    arrowLeftVisible: false,
-    arrowRightVisible: true,
+    arrowLeftActive: false,
+    arrowRightActive: true,
+    ariaLabelLeft: 'Link to previous page, disabled',
+    ariaLabelRight: 'Link to next page',
     pages: [
       {
         pageNo: '1',
@@ -11,23 +12,23 @@ module.exports = {
       },
       {
         pageNo: 2,
-        active: false
+        active: false,
+        ariaLabel: 'Link to page 2'
       },
       {
         pageNo: 3,
-        active: false
+        active: false,
+        ariaLabel: 'Link to page 3'
       },
       {
         pageNo: 4,
-        active: false
+        active: false,
+        ariaLabel: 'Link to page 4'
       },
       {
         pageNo: 5,
-        active: false
-      },
-      {
-        pageNo: 6,
-        active: false
+        active: false,
+        ariaLabel: 'Link to page 5'
       }
     ]
   },
@@ -35,32 +36,35 @@ module.exports = {
     {
       name: 'both arrows',
       context: {
-        arrowLeftVisible: true,
-        arrowRightVisible: true,
+        arrowLeftActive: true,
+        arrowRightActive: true,
+        ariaLabelLeft: 'Link to previous page',
+        ariaLabelRight: 'Link to next page',
         pages: [
           {
             pageNo: 1,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 1'
           },
           {
             pageNo: 2,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 2'
           },
           {
             pageNo: 3,
-            active: true
+            active: true,
+            ariaLabel: 'Link to page 3'
           },
           {
             pageNo: 4,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 4'
           },
           {
             pageNo: 5,
-            active: false
-          },
-          {
-            pageNo: 6,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 5'
           }
         ]
       }
@@ -68,31 +72,33 @@ module.exports = {
     {
       name: 'arrow left',
       context: {
-        arrowLeftVisible: true,
-        arrowRightVisible: false,
+        arrowLeftActive: true,
+        arrowRightActive: false,
+        ariaLabelLeft: 'Link to previous page',
+        ariaLabelRight: 'Link to next page, disabled',
         pages: [
           {
             pageNo: 1,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 1'
           },
           {
             pageNo: 2,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 2'
           },
           {
             pageNo: 3,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 3'
           },
           {
             pageNo: 4,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 4'
           },
           {
             pageNo: 5,
-            active: false
-          },
-          {
-            pageNo: 6,
             active: true
           }
         ]
@@ -101,8 +107,10 @@ module.exports = {
     {
       name: 'more sites',
       context: {
-        arrowLeftVisible: true,
-        arrowRightVisible: true,
+        arrowLeftActive: true,
+        arrowRightActive: true,
+        ariaLabelLeft: 'Link to previous page',
+        ariaLabelRight: 'Link to next page',
         pages: [
           {
             pageNo: 6,
@@ -110,11 +118,8 @@ module.exports = {
           },
           {
             pageNo: 7,
-            active: false
-          },
-          {
-            pageNo: 8,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 7'
           },
           {
             pageNo: '...',
@@ -122,11 +127,13 @@ module.exports = {
           },
           {
             pageNo: 17,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 17'
           },
           {
             pageNo: 18,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 18'
           }
         ]
       }
@@ -134,8 +141,10 @@ module.exports = {
     {
       name: 'arrow right',
       context: {
-        arrowLeftVisible: true,
-        arrowRightVisible: true,
+        arrowLeftActive: true,
+        arrowRightActive: true,
+        ariaLabelLeft: 'Link to previous page',
+        ariaLabelRight: 'Link to next page',
         pages: [
           {
             pageNo: 13,
@@ -143,19 +152,23 @@ module.exports = {
           },
           {
             pageNo: 14,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 14'
           },
           {
             pageNo: 15,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 15'
           },
           {
             pageNo: 16,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 16'
           },
           {
             pageNo: 17,
-            active: false
+            active: false,
+            ariaLabel: 'Link to page 17'
           }
         ]
       }