+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Slide module
+ - Arrow-Right icon.
+ - home page view
+ - variant for list - with components as content
+
+### Changed
+ - Pager element styling that fits alpaca design.
 
 ## [0.4.1] - 13.11.2017
 ### Fixed